+++ test/unit/package_admin/create_template_dict.py
@@ -22,8 +22,8 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__