+++ test/unit/package_admin/get_latest_kernel.py
@@ -21,13 +21,13 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class Package(object):
+class Package(object):                                  # pylint:disable=R0903
 
     """Class:  Package