+++ test/unit/package_admin/kernel_check.py
@@ -22,13 +22,13 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class Package(object):
+class Package(object):                                  # pylint:disable=R0903
 
     """Class:  Package
 
@@ -68,7 +68,7 @@ class Package(object):
         return 0
 
 
-class Dnf(object):
+class Dnf(object):                                      # pylint:disable=R0903
 
     """Class:  Dnf