+++ test/unit/package_admin/kernel_run.py
@@ -23,8 +23,8 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -56,7 +56,7 @@ class ArgParser(object):
         self.args_array = {"-i": "Database_Name:Table_Name"}
 
 
-class Dnf(object):
+class Dnf(object):                                      # pylint:disable=R0903
 
     """Class:  Dnf