+++ test/unit/package_admin/kernel_run3.py
@@ -24,13 +24,13 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class ArgParser(object):
+class ArgParser(object):                                # pylint:disable=R0903
 
     """Class:  ArgParser
 
@@ -55,7 +55,7 @@ class ArgParser(object):
         self.args_array = {"-i": "Database_Name:Table_Name"}
 
 
-class Dnf(object):
+class Dnf(object):                                      # pylint:disable=R0903
 
     """Class:  Dnf