+++ test/unit/package_admin/list_ins_pkg.py
@@ -22,8 +22,8 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import package_admin
-import version
+import package_admin                            # pylint:disable=E0401,C0413
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -55,7 +55,7 @@ class ArgParser(object):
         self.args_array = {"-i": "Database_Name:Table_Name"}
 
 
-class Yum(object):
+class Yum(object):                                      # pylint:disable=R0903
 
     """Class:  Yum